ai_content_factory/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
apscheduler>=3.10,<4.0
pydantic>=2.10,<3.0
pydantic-settings>=2.6,<3.0
python-dotenv>=1.0,<2.0
Pillow>=11.1,<12.0
pytest>=8.3,<9.0

